calculateMAxQPrime logic
