+++ src/main/resources/data/agentstats
@@ -1,3 +1,3 @@
 win_percentage=0.55
-loss_count=843
-win_count=1051
+win_count=1053
+loss_count=845
